MHMVR 0.1a
Completely ramped up in quality, control and the UI is a system taht
works. However, there are still graphical and general efficiency issues
that don't cripple the experience but make the handle and fell rough
around the edges. Also, at this point, not all the assets for this
project were made available to me so there are blank walls.